🌱 Split oversized hooks & lib-layer files under max-lines limit
Splits four files flagged with the 'eslint-disable max-lines -- TODO:
split this file (tracked by #15790)' TODO into focused sibling modules,
removing the suppression comment from each:

- web/src/hooks/useTokenUsage.ts: extracted shared type definitions to
  useTokenUsage.types.ts and pure token-budget math helpers
  (getTokenAlertLevel, reconcileUsageBreakdown, getUsagePeriodKey,
  getNextResetDate) to useTokenUsage.math.ts.
- web/src/lib/auth.tsx: extracted the User/AuthContextType types to
  lib/auth/types.ts and JWT/token-cache helpers (getJwtExpiryMs,
  isJWTExpired, showExpiryWarningBanner, getCachedUser, cacheUser) to
  lib/auth/tokenHelpers.ts.
- web/src/hooks/mcp/compute.ts: extracted useNVIDIAOperators to
  hooks/mcp/compute/nvidia.ts and static GPU/node demo data to
  hooks/mcp/compute/demoData.ts.
- web/src/hooks/useStackDiscovery.ts: extracted resource type
  definitions and aggregation/filtering helpers (isLlmdNamespace,
  isLlmdDeployment, buildComponentsFromDeployments,
  mergeStackWithCached, getStackStatus, cache load/save) to
  useStackDiscovery.helpers.ts.

All original import paths keep working via re-exports. No behaviour
change — verified with a full project tsc --noEmit (zero errors), a
targeted eslint pass on all touched files (zero new warnings vs.
baseline), and the existing vitest suites for these four modules
(421 tests, same pass/fail profile as baseline main).

Fixes #21499
Fixes #21605
Fixes #21606
