+++ web/src/hooks/mcp/compute.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines -- TODO: split this file (tracked by #15790) */
 import { useState, useEffect, useCallback, useMemo, useRef } from 'react'
 import { fetchSSE } from '../../lib/sseClient'
 import { reportAgentDataSuccess, isAgentUnavailable } from '../useLocalAgent'
@@ -12,7 +11,11 @@ import { MCP_EXTENDED_TIMEOUT_MS, MCP_HOOK_TIMEOUT_MS, POLL_INTERVAL_FAST_MS, LO
 import { classifyError, type ClusterErrorType } from '../../lib/errorClassifier'
 import { isInClusterMode } from '../useBackendHealth'
 import { getClusterModeBaseUrl, isClusterModeBackend } from '../../lib/cache/fetcherUtils'
-import type { GPUNode, NodeInfo, NVIDIAOperatorStatus } from './types'
+import type { GPUNode, NodeInfo } from './types'
+import { getDemoGPUNodes, getDemoNodes } from './compute/demoData'
+
+export { useNVIDIAOperators } from './compute/nvidia'
+
 
 // GPU fetch retry configuration
 const GPU_FETCH_MAX_RETRIES = 2
@@ -616,199 +619,6 @@ export function useNodes(cluster?: string) {
   // multi-cluster drill-down can explain an empty list with "lacks
   // list-nodes RBAC on cluster X" rather than a generic warning.
   return { nodes, isLoading, error, clusterErrors, refetch }
-}
-
-// Hook to get NVIDIA operator status
-export function useNVIDIAOperators(cluster?: string) {
-  const [operators, setOperators] = useState<NVIDIAOperatorStatus[]>([])
-  const [isLoading, setIsLoading] = useState(true)
-  const [error, setError] = useState<string | null>(null)
-
-  const refetch = useCallback(async () => {
-    setIsLoading(true)
-    try {
-      const params: Record<string, string> = {}
-      if (cluster) params.cluster = cluster
-
-      const agentBaseUrl = getClusterModeBaseUrl()
-      if (!agentBaseUrl) {
-        setOperators([])
-        setError(null)
-        return
-      }
-
-      // Try SSE streaming first
-      const token = await getStoredAuthToken()
-      if ((token && token !== 'demo-token') || isInClusterMode()) {
-        try {
-          const accumulated: NVIDIAOperatorStatus[] = []
-          const result = await fetchSSE<NVIDIAOperatorStatus>({
-            url: `${agentBaseUrl}/nvidia-operators/stream`,
-            params,
-            itemsKey: 'operators',
-            onClusterData: (_clusterName, items) => {
-              accumulated.push(...items)
-              setOperators([...accumulated])
-              setIsLoading(false)
-            },
-          })
-          setOperators(result)
-          setError(null)
-          setIsLoading(false)
-          return
-        } catch {
-          // SSE failed, fall through to REST
-        }
-      }
-
-      // REST fallback
-      const urlParams = new URLSearchParams()
-      if (cluster) urlParams.append('cluster', cluster)
-      const resp = await agentFetch(`${agentBaseUrl}/nvidia-operators?${urlParams}`)
-      if (!resp.ok) throw new Error(`HTTP ${resp.status}`)
-      const data = await resp.json()
-      if (data.operators) {
-        setOperators(data.operators)
-      } else if (data.operator) {
-        setOperators([data.operator])
-      } else {
-        setOperators([])
-      }
-      setError(null)
-    } catch {
-      setError(null)
-      setOperators([])
-    } finally {
-      setIsLoading(false)
-    }
-  }, [cluster])
-
-  useEffect(() => {
-    refetch()
-  }, [refetch])
-
-  return { operators, isLoading, error, refetch }
-}
-
-// Demo data functions (not exported)
-
-function getDemoGPUNodes(): GPUNode[] {
-  return [
-    // vllm-gpu-cluster - Large GPU cluster for AI/ML workloads
-    { name: 'gpu-node-1', cluster: 'vllm-gpu-cluster', gpuType: 'NVIDIA A100', gpuCount: 8, gpuAllocated: 6, acceleratorType: 'GPU' },
-    { name: 'gpu-node-2', cluster: 'vllm-gpu-cluster', gpuType: 'NVIDIA A100', gpuCount: 8, gpuAllocated: 8, acceleratorType: 'GPU' },
-    { name: 'gpu-node-3', cluster: 'vllm-gpu-cluster', gpuType: 'NVIDIA A100', gpuCount: 8, gpuAllocated: 4, acceleratorType: 'GPU' },
-    { name: 'gpu-node-4', cluster: 'vllm-gpu-cluster', gpuType: 'NVIDIA H100', gpuCount: 8, gpuAllocated: 7, acceleratorType: 'GPU' },
-    // EKS - Production ML inference
-    { name: 'eks-gpu-1', cluster: 'eks-prod-us-east-1', gpuType: 'NVIDIA A10G', gpuCount: 4, gpuAllocated: 3, acceleratorType: 'GPU' },
-    { name: 'eks-gpu-2', cluster: 'eks-prod-us-east-1', gpuType: 'NVIDIA A10G', gpuCount: 4, gpuAllocated: 4, acceleratorType: 'GPU' },
-    // GKE - Training workloads with GPUs and TPUs
-    { name: 'gke-gpu-pool-1', cluster: 'gke-staging', gpuType: 'NVIDIA T4', gpuCount: 2, gpuAllocated: 1, acceleratorType: 'GPU' },
-    { name: 'gke-gpu-pool-2', cluster: 'gke-staging', gpuType: 'NVIDIA T4', gpuCount: 2, gpuAllocated: 2, acceleratorType: 'GPU' },
-    // GKE - TPU nodes (Google Cloud)
-    { name: 'gke-tpu-node-1', cluster: 'gke-staging', gpuType: 'Google TPU v4', gpuCount: 4, gpuAllocated: 3, acceleratorType: 'TPU', manufacturer: 'Google' },
-    { name: 'gke-tpu-node-2', cluster: 'gke-staging', gpuType: 'Google TPU v5e', gpuCount: 8, gpuAllocated: 6, acceleratorType: 'TPU', manufacturer: 'Google' },
-    // AKS - Dev/test GPUs
-    { name: 'aks-gpu-node', cluster: 'aks-dev-westeu', gpuType: 'NVIDIA V100', gpuCount: 2, gpuAllocated: 1, acceleratorType: 'GPU' },
-    // OpenShift - Enterprise ML
-    { name: 'ocp-gpu-worker-1', cluster: 'openshift-prod', gpuType: 'NVIDIA A100', gpuCount: 4, gpuAllocated: 4, acceleratorType: 'GPU' },
-    { name: 'ocp-gpu-worker-2', cluster: 'openshift-prod', gpuType: 'NVIDIA A100', gpuCount: 4, gpuAllocated: 2, acceleratorType: 'GPU' },
-    // Intel Gaudi (AI accelerator, classified as GPU)
-    { name: 'gaudi-node-1', cluster: 'openshift-prod', gpuType: 'Intel Gaudi2', gpuCount: 8, gpuAllocated: 6, acceleratorType: 'GPU', manufacturer: 'Intel' },
-    // IBM AIU nodes (on OCI cluster)
-    { name: 'oci-aiu-node-1', cluster: 'oci-oke-phoenix', gpuType: 'IBM AIU', gpuCount: 4, gpuAllocated: 3, acceleratorType: 'AIU', manufacturer: 'IBM' },
-    { name: 'oci-aiu-node-2', cluster: 'oci-oke-phoenix', gpuType: 'IBM AIU', gpuCount: 4, gpuAllocated: 2, acceleratorType: 'AIU', manufacturer: 'IBM' },
-    // Intel XPU nodes (on AKS cluster)
-    { name: 'aks-xpu-node-1', cluster: 'aks-dev-westeu', gpuType: 'Intel Data Center GPU Max', gpuCount: 4, gpuAllocated: 3, acceleratorType: 'XPU', manufacturer: 'Intel' },
-    { name: 'aks-xpu-node-2', cluster: 'aks-dev-westeu', gpuType: 'Intel Data Center GPU Flex', gpuCount: 8, gpuAllocated: 5, acceleratorType: 'XPU', manufacturer: 'Intel' },
-    // OCI - Oracle GPU shapes
-    { name: 'oke-gpu-node', cluster: 'oci-oke-phoenix', gpuType: 'NVIDIA A10', gpuCount: 4, gpuAllocated: 3, acceleratorType: 'GPU' },
-    // Alibaba - China region ML
-    { name: 'ack-gpu-worker', cluster: 'alibaba-ack-shanghai', gpuType: 'NVIDIA V100', gpuCount: 8, gpuAllocated: 6, acceleratorType: 'GPU' },
-    // Rancher - Managed GPU pool
-    { name: 'rancher-gpu-1', cluster: 'rancher-mgmt', gpuType: 'NVIDIA T4', gpuCount: 2, gpuAllocated: 1, acceleratorType: 'GPU' },
-  ]
-}
-
-function getDemoNodes(): NodeInfo[] {
-  return [
-    {
-      name: 'node-1',
-      cluster: 'prod-east',
-      status: 'Ready',
-      roles: ['control-plane', 'master'],
-      internalIP: '10.0.1.10',
-      kubeletVersion: 'v1.28.4',
-      containerRuntime: 'containerd://1.6.24',
-      os: 'Ubuntu 22.04.3 LTS',
-      architecture: 'amd64',
-      cpuCapacity: '8',
-      memoryCapacity: '32Gi',
-      storageCapacity: '200Gi',
-      podCapacity: '110',
-      conditions: [{ type: 'Ready', status: 'True', reason: 'KubeletReady', message: 'kubelet is posting ready status' }],
-      labels: { 'node-role.kubernetes.io/control-plane': '' },
-      taints: ['node-role.kubernetes.io/control-plane:NoSchedule'],
-      age: '45d',
-      unschedulable: false,
-    },
-    {
-      name: 'node-2',
-      cluster: 'prod-east',
-      status: 'Ready',
-      roles: ['worker'],
-      internalIP: '10.0.1.11',
-      kubeletVersion: 'v1.28.4',
-      containerRuntime: 'containerd://1.6.24',
-      os: 'Ubuntu 22.04.3 LTS',
-      architecture: 'amd64',
-      cpuCapacity: '16',
-      memoryCapacity: '64Gi',
-      storageCapacity: '500Gi',
-      podCapacity: '110',
-      conditions: [{ type: 'Ready', status: 'True', reason: 'KubeletReady', message: 'kubelet is posting ready status' }],
-      labels: { 'node.kubernetes.io/instance-type': 'm5.4xlarge' },
-      age: '45d',
-      unschedulable: false,
-    },
-    {
-      name: 'gpu-node-1',
-      cluster: 'vllm-d',
-      status: 'Ready',
-      roles: ['worker'],
-      internalIP: '10.0.2.20',
-      kubeletVersion: 'v1.28.4',
-      containerRuntime: 'containerd://1.6.24',
-      os: 'Ubuntu 22.04.3 LTS',
-      architecture: 'amd64',
-      cpuCapacity: '32',
-      memoryCapacity: '128Gi',
-      storageCapacity: '1Ti',
-      podCapacity: '110',
-      conditions: [{ type: 'Ready', status: 'True', reason: 'KubeletReady', message: 'kubelet is posting ready status' }],
-      labels: { 'nvidia.com/gpu': 'true', 'node.kubernetes.io/instance-type': 'p3.8xlarge' },
-      age: '30d',
-      unschedulable: false,
-    },
-    {
-      name: 'kind-control-plane',
-      cluster: 'kind-local',
-      status: 'Ready',
-      roles: ['control-plane'],
-      internalIP: '172.18.0.2',
-      kubeletVersion: 'v1.27.3',
-      containerRuntime: 'containerd://1.7.1',
-      os: 'Ubuntu 22.04.2 LTS',
-      architecture: 'amd64',
-      cpuCapacity: '4',
-      memoryCapacity: '8Gi',
-      storageCapacity: '50Gi',
-      podCapacity: '110',
-      conditions: [{ type: 'Ready', status: 'True', reason: 'KubeletReady', message: 'kubelet is posting ready status' }],
-      age: '7d',
-      unschedulable: false,
-    },
-  ]
 }
 
 export const __computeTestables = {
